// src/core/types/inputs/technician.ts
import type { Specialty } from '@/core/types/enums/specialty';
import type { TechnicianId } from '@/core/types/primitives/ids';

export interface TechnicianInput {
    id: TechnicianId;
    name: string;
    speciality?: Specialty;
    specialty?: Specialty[];
    startTime: string;
    endTime: string;
    efficiency?: number;
    lunchBreak?: string;
}